[JAVA-41879] Moving some article links on Github - core-java-regex (#18132)
