+++ core-java-modules/core-java-regex-2/README.md
@@ -6,8 +6,6 @@
 - [How to Use Regular Expressions to Replace Tokens in Strings in Java](https://www.baeldung.com/java-regex-token-replacement)
 - [Creating a Java Array from Regular Expression Matches](https://www.baeldung.com/java-array-regex-matches)
 - [Getting the Text That Follows After the Regex Match in Java](https://www.baeldung.com/java-regex-text-after-match)
-- [Regular Expression: \z vs \Z Anchors in Java](https://www.baeldung.com/java-regular-expression-z-vs-z-anchors)
 - [Extract Text Between Square Brackets](https://www.baeldung.com/java-get-content-between-square-brackets)
 - [Get the Indexes of Regex Pattern Matches in Java](https://www.baeldung.com/java-indexes-regex-pattern-matches)
-- [Check if a String Is Strictly Alphanumeric With Java](https://www.baeldung.com/java-check-string-contains-only-letters-numbers)
-- More articles: [[<-- prev]](/core-java-modules/core-java-regex)
+- More articles: [[<-- prev]](../core-java-regex)[[next -->]](../core-java-regex-3)
