+++ core-java-modules/core-java-regex-3/src/test/java/com/baeldung/patternreuse/PatternUnitTest.java
@@ -1,6 +1,8 @@
 package com.baeldung.patternreuse;
 
-import org.junit.jupiter.api.Test;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertTrue;
 
 import java.util.Arrays;
 import java.util.List;
@@ -10,7 +12,7 @@ import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-import static org.junit.Assert.*;
+import org.junit.jupiter.api.Test;
 
 public class PatternUnitTest {
