+++ core-java-modules/core-java-regex-4/README.md
@@ -1,0 +1,5 @@
+### Relevant Articles: 
+
+- [How to Count the Number of Matches for a Regex?](https://www.baeldung.com/java-count-regex-matches)
+- [Find All Numbers in a String in Java](https://www.baeldung.com/java-find-numbers-in-string)
+- More articles: [[<-- prev]](/core-java-modules/core-java-regex-3)
