+++ core-java-modules/core-java-regex/README.md
@@ -4,14 +4,10 @@
 
 ### Relevant Articles: 
 
-- [An Overview of Regular Expressions Performance in Java](https://www.baeldung.com/java-regex-performance)
 - [A Guide To Java Regular Expressions API](http://www.baeldung.com/regular-expressions-java)
 - [Guide to Escaping Characters in Java RegExps](http://www.baeldung.com/java-regexp-escape-char)
-- [Pre-compile Regex Patterns Into Pattern Objects](https://www.baeldung.com/java-regex-pre-compile)
 - [Difference Between Java Matcher find() and matches()](https://www.baeldung.com/java-matcher-find-vs-matches)
 - [Regular Expressions \s and \s+ in Java](https://www.baeldung.com/java-regex-s-splus)
 - [Validate Phone Numbers With Java Regex](https://www.baeldung.com/java-regex-validate-phone-numbers)
-- [How to Count the Number of Matches for a Regex?](https://www.baeldung.com/java-count-regex-matches)
-- [Find All Numbers in a String in Java](https://www.baeldung.com/java-find-numbers-in-string)
-- [Understanding the Pattern.quote Method](https://www.baeldung.com/java-pattern-quote)
+- [Check if a String Is Strictly Alphanumeric With Java](https://www.baeldung.com/java-check-string-contains-only-letters-numbers)
 - More articles: [[next -->]](/core-java-modules/core-java-regex-2)
