+++ core-java-modules/core-java-regex/src/main/java/com/baeldung/alphanumeric/AlphanumericPerformanceBenchmark.java
@@ -3,6 +3,7 @@ package com.baeldung.alphanumeric;
 import java.util.concurrent.TimeUnit;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
+
 import org.openjdk.jmh.annotations.Benchmark;
 import org.openjdk.jmh.annotations.BenchmarkMode;
 import org.openjdk.jmh.annotations.Fork;
