+++ core-java-modules/core-java-regex/src/test/java/com/baeldung/alphanumeric/AlphanumericUnitTest.java
@@ -1,6 +1,6 @@
 package com.baeldung.alphanumeric;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
